third_party/blink/web_tests/plugins/mouse-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/plugin.js"></script>
<script>
startAfterLoadAndFinish(function () {
    if (!window.testRunner) {
        document.body.appendChild(document.createTextNode("This test does not work in manual mode."));
        return;
    }

    testRunner.dumpAsText();

    eventSender.mouseMoveTo(0,0);
    eventSender.mouseMoveTo(20,20);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.mouseMoveTo(30,30);
    eventSender.mouseDown();
    eventSender.mouseMoveTo(40,30);
    eventSender.mouseUp();
    eventSender.mouseMoveTo(0,0);
});
</script>
</head>
<body>
<embed type="application/x-blink-test-plugin" width=100 height=100></embed>
<p>Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=11517">bug 11517</a>:
Flash clicks/interactivity not working properly.</p>
</body>
</html>
