third_party/blink/web_tests/plugins/mouse-events-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (12,12)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (12,12)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (22,22)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (32,22)
Test for bug 11517: Flash clicks/interactivity not working properly.
